Crate fasten_clean

Trim and filter reads

Examples

cat testdata/four_lines.fastq | \
  fasten_clean > out.fastq

more options

cat testdata | \
  fasten_clean --min-avg-quality 25 --min-trim-quality 25 \
  > out.fastq
 

Usage

Usage: fasten_clean [-h] [-n INT] [-p] [-v] [--min-length INT] [--min-avg-quality FLOAT] [--min-trim-quality INT]

   Options:
       -h, --help          Print this help menu.
       -n, --numcpus INT   Number of CPUs (default: 1)
       -p, --paired-end    The input reads are interleaved paired-end
       -v, --verbose       Print more status messages
           --min-length INT
                           Minimum length for each read in bp
           --min-avg-quality FLOAT
                           Minimum average quality for each read
           --min-trim-quality INT
                           Trim the edges of each read until a nucleotide of at
                           least X quality is found

Functions

avg_quality

Determine average quality of a qual cigar string, e.g., let q:f32 = avg_quality(“AABC!…”)

clean_entry

Cleans a SE or PE read

main

trim

Trim the ends of reads with low quality